Progress on DFA converter
Sorry for the mess. Currently creates two new arraylists.

The first one is simply a duplicate of all current states in the NFA.

The second is where we will push the new states.

What's currently happening is that first it gets the initial state as the cur_state.

It checks for all transitions at a certain input.

If these transitions create a new "NFA state", we create that state.

In this new state, we add ALL transitions of those transitions.

e.g.: A -> AB therefore
AB must get: A -> AB and B -> (?)

as their own transitions now. (AB -> AB and  AB -> (?) )

This new state (new or not actually) gets added to the second list.

The first list and second list are actually the same states in  terms of reference, but we only get the necessary ones in the second list. The first list is simply f or referencing and grabbing transitions.

Everytime we come across a state we haven't seen in the second list, it gets added to the stateStack (it's actually a queue lol I should rename it)

We then pop this queue once the whole o peration is done and go to the next state.