Added deadState code, finished NFA algo
Deadstate creates a _deadState state since there are no "_" in regex we can be sure that this will be taken without issue.

Upon finding deadState transition, we make the current state transition to the deadstate.

If we come across an existing NFA connection, it will reroute the cur_state's transition to said NFA state instead of a "new NFA state".
e.g.: A -> BC

since BC exists, we won't do (A->B && A->C) instead we now have A->BC

Finally, cur_state will be null if there are no more states in queue.

We create the machine at line 170 with the appropriate details and then simply subject it to the reduction algorithm and return  said machine.